// packages/story-tools/src/type.ts
export type Callback<T = HTMLElement> = (element: T) => void
// ...
export type InjectOptions = {
  type: 'script' | 'link'
  props?: WritableHTMLScriptElement | WritableHTMLLinkElement
}

export type WritableScriptKeys =
  | 'async'
  | 'charset'
  | 'crossOrigin'
  | 'defer'
  | 'event'
  | 'htmlFor'
  | 'integrity'
  | 'noModule'
  | 'referrerPolicy'
  | 'src'
  | 'text'
  | 'type'
export type WritableHTMLScriptElement = Pick<
  HTMLScriptElement,
  WritableScriptKeys
>
export type WritableLinkKeys =
  | 'as'
  | 'charset'
  | 'crossOrigin'
  | 'disabled'
  | 'href'
  | 'hreflang'
  | 'imageSizes'
  | 'rel'
  | 'type'
export type WritableHTMLLinkElement = Pick<HTMLLinkElement, WritableLinkKeys>